+++ lib/modules/navigation/navigation_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: navigation
-// Updated at: 2026-04-02 18:20:19
-// Commit: feat(navigation): add bi-directional event bus handler for navigation
+// Updated at: 2026-04-02 19:02:45
+// Commit: fix(navigation): prevent double tap duplicate action in navigation
 
 class navigationComponentState {
-  final String commitRef = '73';
-  final String lastMessage = 'feat(navigation): add bi-directional event bus handler for navigation';
-  final DateTime timestamp = DateTime.parse('2026-04-02T18:20:19+05:00');
+  final String commitRef = '77';
+  final String lastMessage = 'fix(navigation): prevent double tap duplicate action in navigation';
+  final DateTime timestamp = DateTime.parse('2026-04-02T19:02:45+05:00');
 }